setup add to history in video player screen

+++ app/src/main/java/com/karimsinouh/youtixv2/database/HistoryDAO.kt
@@ -27,7 +27,7 @@ interface HistoryDAO {
     suspend fun exists(videoId:String):Boolean
 
     @Query("UPDATE HistoryItem SET currentMillis =:millis WHERE videoId=:videoId ")
-    suspend fun updateMillis(videoId:String,millis:Long)
+    suspend fun updateMillis(videoId:String,millis:Int)
 
     @Query("DELETE  FROM HistoryItem")
     fun clear()
